+++ src/common/gitProvider/github.ts
@@ -4,7 +4,7 @@ import { GitProviderRoot, PullRequestCommentRef } from "./gitProviderRoot.js";
 import { getCurrentGitBranch, git, uxLog } from "../utils/index.js";
 import { CommonPullRequestInfo, CreatePullRequestRequest, CreatePullRequestResult, PullRequestMessageRequest, PullRequestMessageResult } from "./index.js";
 import { GitHub } from "@actions/github/lib/utils.js";
-import { CONSTANTS, getBannerMarkdownAndLink } from "../../config/index.js";
+import { CONSTANTS, getBannerMarkdownAndLink, getPrCommentBannerMarkdown } from "../../config/index.js";
 import { t } from '../utils/i18n.js';
 import { isJenkins, getJenkinsBranchName, getJenkinsPrNumber, getJenkinsBuildNumber, getJenkinsJobName, getJenkinsJobUrl } from "./jenkinsUtils.js";
 
@@ -372,9 +372,9 @@ export class GithubProvider extends GitProviderRoot {
     const githubJobUrl = await this.getCurrentJobUrl();
     // Build note message
     const messageKey = prMessage.messageKey + "-" + this.workflow + "-" + this.prNumber;
-    let messageBody = `## ${prMessage.title || ""}
+    let messageBody = `${getPrCommentBannerMarkdown(prMessage.bannerKey)}## ${prMessage.title || ""}
 
-${prMessage.message}
+${prMessage.navBlock || ""}${prMessage.message}
 
 _Powered by [sfdx-hardis](${CONSTANTS.DOC_URL_ROOT}) from job [${this.workflow}](${githubJobUrl})_
 
@@ -771,7 +771,7 @@ ${getBannerMarkdownAndLink()}
     const results: PullRequestCommentRef[] = [];
     for (const comment of comments) {
       if (comment?.body?.includes(marker)) {
-        results.push({ prNumber: issueNumber, ref: comment.id, body: comment.body });
+        results.push({ prNumber: issueNumber, ref: comment.id, body: comment.body, url: comment.html_url || '' });
       }
     }
     return results;